+++ R/zzz.R
@@ -14,6 +14,6 @@
     "Q", "Q_c", "SQTBA", "lat_rnd", "lon_rnd", "sigma",
     "hour.inc", "ygrid", "xgrid", "SQTBA_norm",
     "start_year", "end_year", "process.recip.lmo",
-    "process.h", "stability"
+    "process.h", "stability", "y1", "y2"
   ))
 }