+++ NEWS.md
@@ -32,6 +32,8 @@
 
 - Move regression formula off main plot for `polarPlot()` for clarity and label slope as 'm'.
 
+- new function `runRegression()` for extracting 'dilution lines' from air quality and other data. Online manual will be updated with principles and examples. 
+
 ## Bug Fixes
 
 - The order of columns in `importUKAQ()` will remain consistent (metadata, date, pollutants, meteo) regardless of whether `hc` is `TRUE` or `FALSE`.